test(zkemv): deactivate deprecated tests
(1) The functions in the aggregation test are only compiled for the `test_agg()`
when the "prover_verify" feature is enabled.
`#[cfg(feature = "prove_verify")]` was added to the other functions as well.

(2) Since the trace file, which is presumed to have been in this repository
a long time ago, is missing, `integration.rs/estimate_circuit_rows()` is
not available. Therefore, this test was deativate by attaching `#[ignore]`